refactor(agent): move data store impls to a separate subpackage
